nvim: bump
neogit: increase timeout for console as it's just annoying to show that
all the time
dap: revert (some) changes for the python adapter. At it turns out, it
is necessary to configure the adapter for an attach request (even if not
using it directly). At least when testing with dash. The debugger
attaches and so on (just using the launch request), but when accessing
the page in dash (eg, `localhost:XXXX`), the page won't load without it.
Debugpy seems to spawn another process that uses the attach request.

also, add the `pwa-chrome` adapter, as I might be using it in the near
future (and it works!). Will be doing some testing with firefox's
adapter as well, but I read some issues and it looks it doesn't work as
nearly as good as chrome's

tabby: remove nvim logo