nvim!: remove molten-nvim
Notebook setup is too complicated and isn't a great experience.

When it comes to editing jupyter files, the setup with jupytext.nvim is
rather sluggish with large notebooks. Even if that weren't an issue,
it's overall awkward to use. For instance, the display of images and
even pandas' dataframes is lacking and the interaction with the kernel
also gets really slow. When the need to edit notebooks comes again, I'll
give this setup a second chance but I'm not very optimistic.

The non-notebook experience suffers with some of these issues too. But
I barely use it, so currently there's no need to keep molten around.