nvim: once more try treesitter folds over ufo
This time we're using an autocmd to properly enable the folds. And also,
it has a workaround (from Lewis, a core maintainer) that handles the
error when restoring a session (ie, `vim.cmd.normal('zx')`). Overall I
think this makes the usage of folds more robust (not only when restoring
sessions) but that's mostly a "feeling". Ideally, we wouldn't have to
use this workaround (some work is being done on core to avoid it).

Some other limitations from treesitter folds still remain. Particularly,
neovim#30352 handicaps the usage of folds in injected languages, which
sucks when using svelte.