nvim: cleanup autocmds
- split autocmds file in 2 (for now) (a file exclusively for LSP related
  autocmds)
- move autocmds that apply to a single filetype for their given `ftplugin`
file (or at least start to do so). Currently, only the dap-repl file has
been migrated. I would also have migrated the svelte one, but I'm not
sure if it's still necessary (I already have created the file with the
setting (commented out), and if I manage to reproduce the issue, I'll
actually commit the file)
- some autocmds don't need a pattern
- lsp: move gd and grr mappings to telescope config
- lsp: remove code_action keymap (there's a default one now)