nvim: vim.wo[0][0] == setlocal
Except when we have to handle restoring sessions: if so, the "0th"
buffer may not be the current one, for some reason unbeknownst to me.

For now, I've noticed this causes issues with treesitter (i.e.,
sometimes you restore the session and the fold options aren't applied),
but I might make some followup changes in the future.