nvim: minor changes and bump spellcheck
catppuccin: slightly darker comments. Comments used to be a nice color,
but they were broken because they were not following the style guide
(now the default one is just too bright). Might revert if I get used to
the new one.
telescope: there's a new default mapping for LSP, so we just override it
with telescope