nvim(octo): avoid jumping around when reviewing PRs
When diffing files in a review, if a line contains a comment, octo
switches the non current window to a window showing the thread. The
issue is that this disables the diff related fold for the remaining
window, which when coupled with treesitter folds, fully collapses the
folds, which makes you lose track of current line. A diagram:

+-----------------------+
|{          |{          |
| ...       | ...       |
|           |           | -> Let's say there's a comment here (line X)
| ...       | ...       |
|}          |}          |
|           |           |
+-----------------------+
                  ^
                  Current window

When moving the cursor to line X, the thread opens and the view becomes

+-----------------------+
|           |{}         | -> Oh no, the fold collapsed!
| thread    |           |       Now I can't get back to line X, since
|           |           |       moving the cursor, changes the position
|           |           |       to either below or above the fold range
|           |           |
|           |           |
+-----------------------+

The workaround is to simply disable automatically showing threads, which
is shame, since it's a nice functionality

See https://github.com/pwntester/octo.nvim/issues/536