nvim: treesitter folds beat lsp folds any day
Initially I enabled LSP folds because of an issue with Treesitter folds,
which affects Svelte exclusively. The issue hasn't been solved yet, but
I guess it's not a big deal anyway. Other than that, TS folds are
working wonderfully (now that I can use them with sessions). In fact,
the workaround for that is no longer necessary.