nvim!: remove vim-illuminate
Who would win: a plugin or two autocmds?

But in all seriousness, I've started noticing some oddities with
vim-illuminate when editing files with non ASCII characters, which
cannot be reproduced with the default implementation.

That's such a shame, since I avoided using the autocmds because their
behavior is inconsistent: it's possible for two words to be highlighted
if the cursor moves too fast. This will probably be fixed by
neovim/neovim#30841, which is a bit stale now.

In the meantime, while both implementations are faulty, I'd rather just
say goodbye to the plugin.