nvim: use my own, custom statusline
Lualine has some downsides. The repo isn't very actively maintained.
There are over 70 open PRs currently. That's obviously not a big deal by
itself, but it's a contributing factor in dropping it.

The major reason is the fact that it causes some really annoying bugs.
As noted in 8458c80, lualine's refresh causes a really odd interaction
with one-small-step-for-vimkind, where it makes it impossible to step.
As also noted on that commit, this would only happen when using certain
custom components, which made the situation difficult to track down.

Additionally, lualine causes some wacky interactions with neovim itself
(see neovim/neovim#35993).

I'm sure these issues could be handled on lualine's side, but given the
lack of activity on the repo, I decided to look for alternatives. There
are plenty of options, but ultimately, I decided to roll my own.

For starters, a base implementation is not difficult at all. A
reasonable chunk of my lualine config were custom components anyway. The
git related ones are made possible by gitsigns.nvim (otherwise they
would be a pain).

Hopefully, it will require minimal maintenance, so it should be worth it.

There are some quirks with redraws, though. Some of which I haven't
gotten around yet. But they shouldn't be a massive pain to deal with.